refactors enrollment process (again)
